feat(observability): added hook for parameters to enable debugging
Also replaced the test container module to speed up tests.
