+++ dynastore_test.go
@@ -1,21 +1,18 @@
 package dynastore
 
 import (
-	"context"
 	"fmt"
-	"io/ioutil"
-	"strconv"
+	"os"
 	"testing"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
-	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbiface"
-	"github.com/dhui/dktest"
-	"github.com/stretchr/testify/require"
+	"github.com/ory/dockertest"
+	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
 )
 
 const (
@@ -23,379 +20,56 @@ const (
 )
 
 var (
-	opts = dktest.Options{PortRequired: true, ReadyFunc: isReady}
+	dbSvc    *dynamodb.DynamoDB
+	endpoint string
 )
 
-type indexFields struct {
-	Created string `json:"created"`
-}
+func TestMain(m *testing.M) {
+	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr, TimeFormat: time.Kitchen}).With().Stack().Caller().Logger()
 
-func isReady(ctx context.Context, c dktest.ContainerInfo) bool {
-	dbSvc := dynamodb.New(mustSession(c.FirstPort()))
-
-	_, err := dbSvc.ListTablesWithContext(ctx, &dynamodb.ListTablesInput{})
-
-	return err == nil
-}
-
-func Test(t *testing.T) {
-	dktest.Run(t, "amazon/dynamodb-local:latest", opts,
-		func(t *testing.T, c dktest.ContainerInfo) {
-			assert := require.New(t)
-
-			dbSvc := dynamodb.New(mustSession(c.FirstPort()))
-
-			err := ensureVersionTable(dbSvc, "testing-locks")
-			assert.NoError(err)
-
-			dl := &DynaSession{DynamoDBAPI: dbSvc}
-
-			testPutGetDeleteExists(t, dl)
-			testList(t, dl)
-			testListPage(t, dl)
-			testAtomicPut(t, dl)
-			testAtomicPutIndex(t, dl)
-			testAtomicDelete(t, dl)
-		})
-}
-
-func mustSession(hostIP, hostPort string, err error) *session.Session {
+	pool, err := dockertest.NewPool("")
 	if err != nil {
-		panic(err)
+		log.Fatal().Msgf("Could not connect to docker: %s", err)
 	}
 
-	ddbURL := fmt.Sprintf("http://%s:%s", hostIP, hostPort)
+	// pulls an image, creates a container based on it and runs it
+	resource, err := pool.Run("amazon/dynamodb-local", "latest", []string{})
+	if err != nil {
+		log.Fatal().Msgf("Could not start resource: %s", err)
+	}
 
+	endpoint = fmt.Sprintf("http://localhost:%s", resource.GetPort("8000/tcp"))
+
+	// exponential backoff-retry, because the application in the container might not be ready to accept connections yet
+	if err := pool.Retry(func() error {
+		dbSvc = dynamodb.New(session.Must(session.NewSession(mustConfig(endpoint))))
+
+		_, err := dbSvc.ListTables(&dynamodb.ListTablesInput{})
+		if err != nil {
+			log.Info().Msgf("Failed to create dynamodb client: %v", err)
+			return err
+		}
+		log.Printf("%#v\n", endpoint)
+		return nil
+	}); err != nil {
+		log.Fatal().Msgf("Could not connect to docker: %s", err)
+	}
+
+	code := m.Run()
+
+	// You can't defer this because os.Exit doesn't care for defer
+	if err := pool.Purge(resource); err != nil {
+		log.Fatal().Msgf("Could not purge resource: %s", err)
+	}
+
+	os.Exit(code)
+}
+
+func mustConfig(endpoint string) *aws.Config {
 	creds := credentials.NewStaticCredentials("123", "test", "test")
-	return session.Must(session.NewSession(&aws.Config{
+	return &aws.Config{
 		Region:      aws.String(defaultRegion),
-		Endpoint:    aws.String(ddbURL),
+		Endpoint:    aws.String(endpoint),
 		Credentials: creds,
-	}))
-}
-
-func ensureVersionTable(dbSvc dynamodbiface.DynamoDBAPI, tableName string) error {
-	_, err := dbSvc.CreateTable(&dynamodb.CreateTableInput{
-		TableName: aws.String(tableName),
-		KeySchema: []*dynamodb.KeySchemaElement{
-			{AttributeName: aws.String("id"), KeyType: aws.String(dynamodb.KeyTypeHash)},
-			{AttributeName: aws.String("name"), KeyType: aws.String(dynamodb.KeyTypeRange)},
-		},
-		LocalSecondaryIndexes: []*dynamodb.LocalSecondaryIndex{
-			{
-				IndexName: aws.String("idx_created"),
-				KeySchema: []*dynamodb.KeySchemaElement{
-					{AttributeName: aws.String("id"), KeyType: aws.String(dynamodb.KeyTypeHash)},
-					{AttributeName: aws.String("created"), KeyType: aws.String(dynamodb.KeyTypeRange)},
-				},
-				Projection: &dynamodb.Projection{ProjectionType: aws.String(dynamodb.ProjectionTypeAll)},
-			},
-		},
-		AttributeDefinitions: []*dynamodb.AttributeDefinition{
-			{AttributeName: aws.String("id"), AttributeType: aws.String(dynamodb.ScalarAttributeTypeS)},
-			{AttributeName: aws.String("name"), AttributeType: aws.String(dynamodb.ScalarAttributeTypeS)},
-			{AttributeName: aws.String("created"), AttributeType: aws.String(dynamodb.ScalarAttributeTypeS)},
-		},
-		ProvisionedThroughput: &dynamodb.ProvisionedThroughput{
-			ReadCapacityUnits:  aws.Int64(1),
-			WriteCapacityUnits: aws.Int64(1),
-		},
-		SSESpecification: &dynamodb.SSESpecification{
-			Enabled: aws.Bool(true),
-			SSEType: aws.String(dynamodb.SSETypeAes256),
-		},
-	})
-
-	if err != nil {
-		if aerr, ok := err.(awserr.Error); ok {
-			if aerr.Code() == dynamodb.ErrCodeResourceInUseException {
-				return nil
-			}
-		}
-		return err
 	}
-
-	err = dbSvc.WaitUntilTableExists(&dynamodb.DescribeTableInput{
-		TableName: aws.String(tableName),
-	})
-	if err != nil {
-		return err
-	}
-
-	_, err = dbSvc.UpdateTimeToLive(&dynamodb.UpdateTimeToLiveInput{
-		TableName: aws.String(tableName),
-		TimeToLiveSpecification: &dynamodb.TimeToLiveSpecification{
-			AttributeName: aws.String("expires"),
-			Enabled:       aws.Bool(true),
-		},
-	})
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func testPutGetDeleteExists(t *testing.T, dSession *DynaSession) {
-	assert := require.New(t)
-
-	// , tableName: "testing-locks", partition: "agent"
-	kv := dSession.Table("testing-locks").Partition("agent")
-
-	// Get a not exist key should return ErrKeyNotFound
-	_, err := kv.Get("testPutGetDelete_not_exist_key")
-	assert.Equal(ErrKeyNotFound, err)
-
-	data, err := ioutil.ReadFile("fixtures/pr.json")
-	assert.NoError(err)
-
-	value := string(data)
-
-	for _, key := range []string{
-		"testPutGetDeleteExists",
-		"testPutGetDeleteExists/",
-		"testPutGetDeleteExists/testbar/",
-		"testPutGetDeleteExists/testbar/testfoobar",
-	} {
-		t.Run(key, func(t *testing.T) {
-			// Put the key
-			err = kv.Put(key, WriteWithString(value), WriteWithTTL(2*time.Second))
-			assert.NoError(err)
-
-			var pair *KVPair
-
-			// Get should return the value and an incremented index
-			pair, err = kv.Get(key)
-			assert.NoError(err)
-			assert.NotNil(pair)
-			assert.Equal(value, pair.StringValue())
-			assert.NotEqual(0, pair.Expires)
-
-			assert.NotEqual(0, pair.Version)
-
-			var exists bool
-
-			// Exists should return true
-			exists, err = kv.Exists(key)
-			assert.NoError(err)
-			assert.True(exists)
-
-			// Delete the key
-			err = kv.Delete(key)
-			assert.NoError(err)
-
-			// Get should fail
-			pair, err = kv.Get(key)
-			assert.Error(err)
-			assert.Nil(pair)
-			assert.Nil(pair)
-
-			// Exists should return false
-			exists, err = kv.Exists(key)
-			assert.NoError(err)
-			assert.False(exists)
-		})
-	}
-
-	key := "something/withoutExpires"
-
-	// Put the key
-	err = kv.Put(key, WriteWithString(value), WriteWithNoExpires(), WriteWithNoExpires())
-	assert.NoError(err)
-
-	// Get should return the value and an incremented index
-	pair, err := kv.Get(key)
-	assert.NoError(err)
-	assert.NotNil(pair)
-	assert.Equal(value, pair.StringValue())
-	assert.Equal(int64(0), pair.Expires)
-}
-
-func testAtomicPut(t *testing.T, dSession *DynaSession) {
-	assert := require.New(t)
-
-	kv := dSession.Table("testing-locks").Partition("agent")
-
-	t.Run("AtomicPut", func(t *testing.T) {
-		key := "testAtomicPut"
-		value := []byte("world")
-
-		// Put the key
-		err := kv.Put(key, WriteWithBytes(value))
-		assert.NoError(err)
-
-		// Get should return the value and an incremented index
-		pair, err := kv.Get(key)
-		assert.NoError(err)
-		assert.NotNil(pair)
-		assert.Equal(value, pair.BytesValue())
-		assert.NotEqual(0, pair.Version)
-
-		// This CAS should fail: previous exists.
-		success, _, err := kv.AtomicPut(key, WriteWithString("WORLD"))
-		assert.Error(err)
-		assert.False(success)
-
-		// This CAS should succeed
-		success, _, err = kv.AtomicPut(key, WriteWithPreviousKV(pair), WriteWithBytes([]byte("WORLD")))
-		assert.NoError(err)
-		assert.True(success)
-
-		// This CAS should fail, key has wrong index.
-		pair.Version = 6744
-		success, _, err = kv.AtomicPut(key, WriteWithPreviousKV(pair), WriteWithBytes([]byte("WORLDWORLD")))
-		assert.Equal(err, ErrKeyModified)
-		assert.False(success)
-	})
-}
-
-func testAtomicPutIndex(t *testing.T, dSession *DynaSession) {
-	assert := require.New(t)
-
-	kv := dSession.Table("testing-locks").Partition("agent")
-
-	t.Run("AtomicPut", func(t *testing.T) {
-		key := "testAtomicPutIndex"
-		value := []byte("world")
-
-		timeStamp := "20200103T1100Z"
-
-		// Put the key
-		err := kv.Put(key, WriteWithBytes(value), WriteWithFields(map[string]string{
-			"created": timeStamp,
-		}))
-		assert.NoError(err)
-
-		// Get should return the value and an incremented index
-		page, err := kv.ListPage(timeStamp, ReadWithLocalIndex("idx_created", "created"))
-		assert.NoError(err)
-		assert.Equal(1, len(page.Keys))
-
-		idxFields := new(indexFields)
-		err = page.Keys[0].DecodeFields(idxFields)
-		assert.NoError(err)
-		assert.Equal(timeStamp, idxFields.Created)
-	})
-}
-
-func testAtomicDelete(t *testing.T, dSession *DynaSession) {
-	assert := require.New(t)
-
-	kv := dSession.Table("testing-locks").Partition("agent")
-
-	t.Run("AtomicDelete", func(t *testing.T) {
-		key := "testAtomicDelete"
-		value := []byte("world")
-
-		// Put the key
-		err := kv.Put(key, WriteWithBytes(value))
-		assert.NoError(err)
-
-		// Get should return the value and an incremented index
-		pair, err := kv.Get(key)
-		assert.NoError(err)
-		assert.NotNil(pair)
-		assert.Equal(value, pair.BytesValue())
-		assert.NotEqual(0, pair.Version)
-
-		tempIndex := pair.Version
-
-		// AtomicDelete should fail
-		pair.Version = 6744
-		success, err := kv.AtomicDelete(key, pair)
-		assert.Error(err)
-		assert.False(success)
-
-		// AtomicDelete should succeed
-		pair.Version = tempIndex
-		success, err = kv.AtomicDelete(key, pair)
-		assert.NoError(err)
-		assert.True(success)
-
-		// Delete a non-existent key; should fail
-		success, err = kv.AtomicDelete(key, pair)
-		assert.Equal(ErrKeyNotFound, err)
-		assert.False(success)
-	})
-}
-
-func testList(t *testing.T, dSession *DynaSession) {
-	assert := require.New(t)
-
-	kv := dSession.Table("testing-locks").Partition("agent")
-
-	t.Run("List", func(t *testing.T) {
-		childKey := "testList/child"
-		subfolderKey := "testList/subfolder"
-
-		// Put the first child key
-		err := kv.Put(childKey, WriteWithBytes([]byte("first")))
-		assert.NoError(err)
-
-		// Put the second child key which is also a directory
-		err = kv.Put(subfolderKey, WriteWithBytes([]byte("second")))
-		assert.NoError(err)
-
-		// Put child keys under secondKey
-		for i := 1; i <= 3; i++ {
-			key := "testList/subfolder/key" + strconv.Itoa(i)
-			err = kv.Put(key, WriteWithBytes([]byte("value")))
-			assert.NoError(err)
-		}
-
-		// List should work and return five child entries
-		pairs, err := kv.List("testList/subfolder/key")
-		assert.NoError(err)
-		assert.NotNil(pairs)
-		assert.Equal(3, len(pairs))
-	})
-}
-
-func testListPage(t *testing.T, dSession *DynaSession) {
-	assert := require.New(t)
-
-	kv := dSession.Table("testing-locks").Partition("agent")
-
-	t.Run("ListPage", func(t *testing.T) {
-		childKey := "testList/child"
-		subfolderKey := "testList/subfolder"
-
-		// Put the first child key
-		err := kv.Put(childKey, WriteWithBytes([]byte("first")))
-		assert.NoError(err)
-
-		// Put the second child key which is also a directory
-		err = kv.Put(subfolderKey, WriteWithBytes([]byte("second")))
-		assert.NoError(err)
-
-		// Put child keys under secondKey
-		for i := 1; i <= 3; i++ {
-			key := "testList/subfolder/key" + strconv.Itoa(i)
-			err = kv.Put(key, WriteWithBytes([]byte("value")))
-			assert.NoError(err)
-		}
-
-		// List should work and return child entries
-		page, err := kv.ListPage("testList/subfolder/key")
-		assert.NoError(err)
-		assert.NotNil(page)
-		assert.Equal(3, len(page.Keys))
-
-		// List should work and return all child entries
-		page, err = kv.ListPage("")
-		assert.NoError(err)
-		assert.NotNil(page)
-		assert.Equal(6, len(page.Keys))
-
-		// List should work and return all child entries
-		page, err = kv.ListPage("", ReadWithLimit(5))
-		assert.NoError(err)
-		assert.NotNil(page)
-		assert.Equal(5, len(page.Keys))
-
-		page, err = kv.ListPage("", ReadWithStartKey(page.LastKey))
-		assert.NoError(err)
-		assert.NotNil(page)
-		assert.Equal(1, len(page.Keys))
-	})
 }
