+++ go.mod
@@ -3,8 +3,21 @@ module github.com/wolfeidau/dynastore
 go 1.16
 
 require (
+	github.com/Azure/go-ansiterm v0.0.0-20170929234023-d6e3b3328b78 // indirect
+	github.com/Microsoft/go-winio v0.4.16 // indirect
+	github.com/Nvveen/Gotty v0.0.0-20120604004816-cd527374f1e5 // indirect
 	github.com/aws/aws-sdk-go v1.37.25
-	github.com/dhui/dktest v0.3.4
+	github.com/cenkalti/backoff v2.2.1+incompatible // indirect
+	github.com/containerd/continuity v0.1.0 // indirect
+	github.com/docker/go-connections v0.4.0 // indirect
+	github.com/gotestyourself/gotestyourself v2.2.0+incompatible // indirect
+	github.com/lib/pq v1.8.0 // indirect
 	github.com/mr-tron/base58 v1.2.0
+	github.com/opencontainers/image-spec v1.0.1 // indirect
+	github.com/opencontainers/runc v1.0.1 // indirect
+	github.com/ory/dockertest v3.3.5+incompatible
+	github.com/rs/zerolog v1.23.0
 	github.com/stretchr/testify v1.7.0
+	golang.org/x/net v0.0.0-20210119194325-5f4716e94777 // indirect
+	gotest.tools v2.2.0+incompatible // indirect
 )
